requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-multipart
python-jose[cryptography]
passlib
python-dotenv
jinja2
email-validator
pandas
openpyxl
qrcode[pil]
reportlab
Pillow
pytz
requests